split slack tools into multiple files
